fix: the 2026-09-05 audit — 20 defects, dead code, and no more 0.00 km runs
Full read of all 148 Swift files, the suite, and a declaration-vs-usage scan.
Report: https://claude.ai/code/artifact/8a628d7f-0a69-41df-8182-0b7dfed0543f

A workout with no distance is not a workout
  finish() refuses one on every manual path (announcing .runCancelled, since
  silence reads as a save), the import loop skips them, "Save as-is" on a
  zero-distance checkpoint discards instead, and purgeZeroDistanceWorkouts()
  runs every sync to clear the two rows already in the field store. Auto-walk
  is the one exception during finish — it has zero GPS metres until the Health
  backfill runs, then applies its own stricter 100 m floor.

Correctness
- The suite was RED: a release test pinned CFBundleShortVersionString to
  "1.10". Its sibling had the right shape all along — assert a floor, not a
  literal. A suite everyone expects to be red is how a real regression gets in.
- A failed durable save was announced to the user as a successful one: one
  alert served both outcomes, titled "Saved locally" and claiming "the workout
  is kept in Runner and will retry automatically" for a run persisted nowhere,
  and dismissing it tore down the only screen that could still retry. Two
  outcomes, two alerts; the failure path no longer dismisses. This is exactly
  what CRITICAL 5 exists to prevent, surviving in the one place the user reads.
- Bike rides Runner recorded itself never got CO2 — CO2Estimator ran only in
  the import loop, so the identical ride from Bixi showed a number and ours
  showed nothing, in the workout, the lifetime total, the green-impact card
  and Wrapped.
- A HealthKit retry passed six of eleven columns and let the rest default;
  upsertWorkout assigns co2SavedGrams and autoStarted unconditionally, so a
  successful retry silently erased them.
- Runner's own workouts are now re-imported when no local row claims their
  (type, start) — after a reinstall they survive in HealthKit but not in
  SwiftData, so they vanished from History, Routes, records and badges while
  still counting toward points and streaks.
- A zero-second split (one re-acquired fix closing several km boundaries at
  once — the recorder already suppresses the spoken cue for it) became the
  all-time "Fastest 1 km", rendered as a blank "—" forever, and fired a bogus
  achievement. Splits below 120 s/km are rejected, as is any 5 km window
  containing one.
- SplitStats renumbered every split after a dropped one: km 4 was labelled
  "Km 3", in the list, the chart axis and the "Fastest km" highlight.
- The Routes map never framed the routes. `region` came from `routed`, which
  is empty at first render, and Map(initialPosition:) is honoured once — so
  the camera always fell through to fittingRegion's literal fallback,
  invisible only because that fallback is Montreal.
- The Wrapped banner's dismiss button called markSeen and nothing else, which
  SwiftUI does not observe: the banner stayed on screen.
- "Best run" in the weekly recap was the longest workout of any type. Renamed
  BestRun -> BestEffort; the recap spans every type and so does this row.
- The activity hub's weekly chart filtered weekly buckets by a DAILY cutoff,
  so "Week" drew one bar. Ranges are 12 wk / 1 yr now.
- Clamping a goal skipped persisting it: Swift does not re-enter an observer
  for an assignment made inside it, so the early return skipped the write.
- reset() now clears lastFinishedSnapshot in start(), not reset() (finish()
  sets it and then resets); LocationFilter reads the injected clock.

Performance — the same hoisting the perf commit applied elsewhere
- Today rebuilt the entire Monthly Wrapped on every body pass (2x allBadges,
  12x typeRecords, a per-day heat strip) purely to decide whether to draw a
  banner already dismissed. isSeen is checked first now.
- TrophyRoomView read allBadges as a computed property five times per render;
  HistoryView called it twice for one answer. A computed property is not a
  cache. nextMilestone(from:hasHistory:) lets a caller reuse the pass.
- RoutesView caches decoded routes across rebuilds instead of re-decoding
  every blob on each filter tap; ActivityDetailView derives once.

Dead code and consistency
- Removed the Milestone ladder (superseded by TrophyMath badges and reachable
  from no view), MotionGate.vetoesStart, InsightsMath.chartWindowWeeks, and
  five struct fields computed but never read. Kept allWorkouts()/profileCount()
  and documented them as the test seams they are. pointsPrevious stays: three
  tests use it to pin the shifted previous-week window, which the delta
  fraction alone cannot distinguish from an empty week.
- Localized format strings now use %lld/Int64 throughout, in code and catalog
  (edited textually — never round-trip that file).
- Shared test doubles moved to Fixtures.swift: FakeLocationProvider lived
  inside WorkoutRecorderTests while four other files depended on it, and
  AnnouncementSpy existed six times byte-identical.
- SyncCoordinator's batched import rolls the context back on failure instead
  of leaving inserted objects for the next unrelated save() to commit.
- ProfileView also saves when the scene leaves .active — onDisappear alone
  lost weight edits to a force-quit, and weight gates every calorie number.
- docs/REVIEW-LOG.md indexes the ~90 comments citing CRITICAL/IMPORTANT/Task/
  fix-wave labels that resolved to nothing in the repo. SPEC.md refreshed to
  1.11 (24) with five new invariants.

413 tests green, no warnings. 1.11 (24).

+++ Runner/Features/History/ActivitySpecificBlock.swift
@@ -28,7 +28,7 @@ struct ActivitySpecificBlock: View {
         let yearByKind = Dictionary(uniqueKeysWithValues: thisYear.map { ($0.kind, $0) })
 
         return VStack(alignment: .leading, spacing: 10) {
-            MicroLabel(text: String(format: String(localized: "%d vs all-time"), year))
+            MicroLabel(text: String(format: String(localized: "%lld vs all-time"), Int64(year)))
             if allTime.isEmpty {
                 notEnoughData
             } else {
@@ -88,13 +88,13 @@ struct ActivitySpecificBlock: View {
                 notEnoughData
             } else {
                 LazyVGrid(columns: [GridItem(.flexible()), GridItem(.flexible())], spacing: 10) {
-                    StatTile(label: String(format: String(localized: "Rides in %d"), review.year),
+                    StatTile(label: String(format: String(localized: "Rides in %lld"), Int64(review.year)),
                              value: "\(review.sessions.current)",
                              accent: type.accent)
                     StatTile(label: String(localized: "Rides lifetime"),
                              value: "\(stats.sessions)",
                              accent: .white)
-                    StatTile(label: String(format: String(localized: "CO₂ avoided in %d"), review.year),
+                    StatTile(label: String(format: String(localized: "CO₂ avoided in %lld"), Int64(review.year)),
                              value: Format.co2(grams: review.co2SavedGrams.current),
                              accent: .rLime)
                     StatTile(label: String(localized: "CO₂ avoided lifetime"),
@@ -126,13 +126,13 @@ struct ActivitySpecificBlock: View {
                 notEnoughData
             } else {
                 LazyVGrid(columns: [GridItem(.flexible()), GridItem(.flexible())], spacing: 10) {
-                    StatTile(label: String(format: String(localized: "Hours in %d"), review.year),
+                    StatTile(label: String(format: String(localized: "Hours in %lld"), Int64(review.year)),
                              value: hours(review.movingSeconds.current),
                              accent: type.accent)
                     StatTile(label: String(localized: "Hours lifetime"),
                              value: hours(stats.totalMovingSeconds),
                              accent: .white)
-                    StatTile(label: String(format: String(localized: "Distance in %d"), review.year),
+                    StatTile(label: String(format: String(localized: "Distance in %lld"), Int64(review.year)),
                              value: Format.km(review.distanceMeters.current),
                              accent: .rLime)
                     StatTile(label: String(localized: "Distance lifetime"),
